+++ src/com/zh/web/action/OrderDeliveryAction.java
@@ -57,7 +57,7 @@ public class OrderDeliveryAction extends BaseAction {
 
 	public String editor() throws Exception {
 		LOGGER.debug("editor()");
-		Integer id = this.processingSingleModel.getId();
+		Long id = this.processingSingleModel.getId();
 
 		if (null == id) {
 			throw new ParameterException("任务单号不允许为空!");
@@ -66,7 +66,7 @@ public class OrderDeliveryAction extends BaseAction {
 		// 查询信息
 		LOGGER.debug("editor ProductionTask id " + id);
 		ProcessingSinglePrimary processingSinglePrimary = this.processingSingleModel.getProcessingSinglePrimary();
-		processingSinglePrimary.setId(Integer.valueOf(id));
+		processingSinglePrimary.setId(Long.valueOf(id));
 		ProcessingSinglePrimary reult = processingSinglePrimaryService.query(processingSinglePrimary);
 		this.processingSingleModel.setProcessingSinglePrimary(reult);
